New blog post, let's go!
Finally, getting around to updating the blog and adding new content. This piece is about the changes since last year in AI assisted development. It's obviously been a huge, huge year. And so just wanted to kind of reflect on how that's worked for me. And that's it. Quick update.

I also refreshed all the dependencies as well.